Minor edits in the introduction chapter
While going through the introduction on the airplane, I found some minor
points that should be corrected to make it easier to read.

Also, following the page I am maintaining on reproducible science
(akhlaghi.org/reproducible-science.html), in the "Science and its tools"
section, I also added the analogy with good coding style to a good English
style which is mandatory in publishing any Scientific result.