Printing column information in standard output when requested
Until now it was implicitly assumed that when printing to standard output,
the column values need to be piped into another program for further
processing (there was an option to have a text table output). However, some
times we just want to inspect the contents of the columns on the
command-line and not necessarily pipe it into another program. In such
cases, the column information/meta-data can be very helpful.

The library's `gal_txt_write' function now accepts a new `colinfoinstdout'
argument for such cases. Ideally, it is possible to see if the program's
output is being piped or simply printed. In this case, we could use such a
check and print the column meta-data or not. But I haven't yet fully learnt
how to implement piping features within a C program with Argp.

So until we are able to implement an internal check of piping, the Table
program has a new option (`--colinfoinstdout'). When called (and the output
is to be printed in the standard output), it will print the column
meta-data to the standard output.

This was suggested by Leindert Boogaard.