Corrected link in Synchronizing section of book
Just before the 0.6 release, the script in the old "Configure and build in
RAM" section was significantly improved and fully described in a new
section called "Separate build and source directories". But the
"Synchronizing" section was still pointing the readers to the "Configure
and build in RAM" section. It now points to the new section.