MakeCatalog's --checkupperlimit table saved in proper directory
When called with `--checkupperlimit', MakeCatalog will build a table of all
the positions and values used in the upper-limit estimation. If no output
file is specified this is fine (the table is stored in the running
directory). When the user has asked for the output in a different
directory, the table must also be built there. But until now, this wouldn't
happen and it is built in the current directory.

With this commit, when the user specifies an output name, we will
temporarily set "p->cp.keepinputdir" to `1' and reset it afterwards.

This fixes bug #54186.