Flag initialized every time in MakeCatalog's upperlimit
To speed things up and not take too much memory and CPU resources, we ask
the sigma-clipping of the values measured during the upper-limit
calculations to be done in place. But we weren't actually initializing the
sorting flag (necessary for the current implementation of sigma-clipping).
Therefore, the sorting flag would be set for the random values of the first
object in a thread, and the flag would persit for all the next objects
(which were written in the same data container). This would create
un-reliable values.

With this commit, we initialize the sorting flag to zero before starting
each object's upper-limit magnitude measurement. In the process some
cleaning up was also done in other parts of the code for easy readability.

This fixes bug #54188.