Minor edits in quick start and package manager sections of book
Some small edits were made in these two sections to make them more useful
and easier to read. In particular the discussion on choosing Lzip over Gzip
was moved to a footnote, since it wasn't directly relevant to someone who
just wants to download and install the Gnuastro.

Also, a link to the dependency installation with package managers was also
added in "Quick start". Finally, the package managers were put before the
operating system names toin the dependency installation in package managers
to be easier to notice (since some operating systems share a package
manager).