Convenience/extra features in tests/during-dev.sh
Until now, this testing script (during development) would continue if there
was a problem. This was very inconvenient (makes it harder to find a
problem). With this commit, a `set -e' has been added to it from its
processing step so it aborts in case of any problem.

When calling this script, the utility name must not have the `ast'
prefix. To avoid errors in copying the files later, a test has been added
that will abort the script if the utility name starts with an `ast'.

The checks on necessary variables were typeset more clearly (things under
each other) to be more easily readable.