Corrected type of number counter in collapse-mean library
The library's `gal_dimension_collapse_mean' function, uses the same
counting pre-processor macros as `gal_dimension_collapse_sum'. But when
calculating the sum, the actual number of points isn't important, only the
fact that atleast one point was present there is important. So to speed up
the sum operator and allow it to consume less memory, the counter dataset
there has an unsigned 8-bit integer type. But in the mean, the number is
important, so it has a signed 32-bit integer type. Since they have the same
name, the macros will work on both.

But the array pointers in `gal_dimension_collapse_mean', were mistakenly
`int8_t' type (similar to the sum library function, since it was initially
just a copy of that). With this commit the types of the pointers are
corrected to be as described above.