Size of reference corrected after collapse in Arithmetic
When reading each operand into memory, Arithmetic uses a reference dataset
to make sure the next given operand has the same size. But the new
collapsing operators change the necessary size for the next steps. So the
reference needs to be corrected after the collapse.

We were already correcting the reference's WCS, but hadn't corrected its
`ndim' and `dsize'. These two are now also corrected with this commit.