Gnuastro's Match program instead of paste in tutorial
To estimate the colors, until now we were using text catalogs and "pasting"
the columns using the command-line's `paste' command. But after going
through the tutorial in the "4th Indo-French Astronomy School" (were many
bugs in the tutorial were found and fixed thanks to the very active
participants), I thought this problem (of finding colors) would be the
perfect situation to also discuss the Match program and how useful it is in
mixing different catalogs.

Therefore, with this commit, the second part of the tutorial (from the
merging of the columns into one catalog) was modified to use Match instead
of `paste'.