Installation of dependencies with YUM described in book
Many astronomers use Red Hat-based OSs, so after a test installation of
Fedora on a virtual machine, the necessary command to install all of
Gnuastro's dependencies with YUM has been added to the "Dependencies from
package managers" section of the book.

Also, the description of Debian-based OSs has been edited to be more clear.