Check for C++ executable to avoid make check crash
According to the Autoconf manual: "if none of the [AC_PROG_CXX] checks
succeed, then as a last resort [it will] set `CXX' to `g++'". Therefore,
when a C++ compiler doesn't exist, when trying to build the
`versioncxx.cpp' test, `make check' will fail! I found this in a fresh
install of Fedora 28 (in Live USB), where `g++' doesn't exist and `make
check' failed.

To fix the problem, a check has been added in `configure.ac' and if the C++
compiler doesn't exist, then the `versioncxx' test will not be executed and
completely ignored!

This fixes bug #54285.