BuildProgram's configuration file not in tarball
For some reason, we hadn't corrected the `dist_' prefix from BuildProgram's
configuration file in its `Makefile.am'. Therefore the configuration file
was being distributed in the tarball and the user's requested `CPPFLAGS'
and `LDFLAGS' were not actually being written into it (it is already built
from the packager's values for these options!!!).

With this commit the `dist_' prefix has been changed to `nodist_' to fix
the problem. But we also need the configuration file during a check. So it
is now also given to `check_DATA'.