Checking contents of checkuplim values instead of pointer
In MakeCatalog's `--checkuplim' option, we store the object (and possible)
clump ID into a two element `p->checkuplim' array. Back when it was
initially defined, the space for this array was only allocated if the user
called it. But very soon afterwards, we converted it into a static array
(initiated with blank values). So to see if the user has called this
option, must check the values: the array will always be non-NULL.

However, in one instance (in `bin/mkcatalog/ui.c'), we were still simply
checking the pointer value not being NULL! This has been corrected with
this commit and we are now checking the contents.

This issue came up in a compiler warning, which was reported by Jenny
Sorce.