Reasonable Match output when there isn't any match
Until now, When there isn't any match, no output would be made, even when
the `--nomatched' option is called. But in the latter case, the user
expects to see an output containing the rows that didn't match (the full
contents of both catalogs).

This is not the expected behavior. With this commit, when there is no match
(and the matched columns are requested), an output will be made and will
contain zero rows. When the `--notmatched' column is requested, the full
input datasets will be put in the two columns.

As part of this check, I also found out that that the
`gal_dimension_is_different' cannot handle the cases when the size is zero
(and `dsize' might be NULL). So this function was also modified to deal
with such cases.

This bug was found and reported by Dmitry Oparin.

This fixes bug #54297.