Table can print column contents with no rows
When there are no rows in the table, until now, Table would crash. It would
complain about not being able to allocate a zero-sized array. With this
commit, this has been fixed by adding checks on the number of rows when
reading FITS and plain text files.

This fixes bug #54298.