Better explanation on some installation issues in book
Some of the installation parts of the book were edited to be more clear and
useful. In particular, some of the major known-issues problems are now
explained in cartouches on top of the package manager commands (so the
cartouche under `yum' was removed) and Lzip has been added to the
list. Also, for RedHat-based distributions, `yum' was replaced by `dnf'
(which is the new version of YUM).