Check for the existence of CFITSIO's fits_is_reentrant
CFITSIO's `fits_is_reentrant' function is only available after version 3.30
(April 2012). So if the system has an older version, Gnuastro's build will
crash. A configuration check has been added in `configure.ac' to do the
check and not call the function if its not present.

Also, when multiple threads are called, but CFITSIO wasn't configured to be
run in multiple threads, Crop won't abort with an error, but the number of
threads will be set to 1.

Finally in `tests/mkprof/mosaic1.sh', we were running MakeProfiles
separately from the `mv' command. So even if MakeProfiles crashed, the test
would succeed, causing really hard to interpret situtations!!!

This bug was reported by Oryna Ivashtenko.

This fixes bug #54312.