Configure option for make check with Valgrind
When a segmentation fault occurs on a user's program during `make check',
until now, it was necessary to ask the user to manually make changes in the
source (which is really not a comfortable operation for the users who do
not necessarily know about debugging or the test shell scripts).

Therefore with this commit, a `--enable-check-with-valgrind' option was
added to the configure script. This option will put the compilation options
to debugging mode, check the existance of Valgrind on the user's system and
run all the tests within Valgrind. This will include Valgrind's output in
the `tests/test-suite.log' file, which is easy to transfer by email for
example and will contain information about any crash.

A corresponding `--valgrind' (or `-v') option was also added to the
`developer-build' script that will use this new configure option.