Crop's checkcenter units now depend on coordinate mode
Except `--checkcenter', the units of the sizes and coordinates that are
input into Crop are based on the `--mode' option. This would cause
confusions.

With this commit, the units of the value given to `--checkcenter' is also
defined by the `--mode' option. Also, the hard constraint (aborting the
program, when the given value was even) has been removed. When cropping in
image mode, if `--checkcenter' is an even number, it will be incremented by
one so the used value will be odd. To clarify this point, in the verbose
output, we now also print the value to this option.

This task was suggested by Juan Molina Tobar.