Upperlimit termination criteria is num failed since last successful
Until now, the termination criteria in randomly placing a profile over a
dataset was a fixed multiple of the total number of requested positions. To
be more efficient in identifying objects that simply cannot fit in any
undetected regions and allow more for those that can, but just need more
time, the criteria was changed as described below.

We now terminate the random positionings when the number of failed attempts
since the last successful attempt reaches a certain multiple of the total
requested number (which is smaller than the pervious total number of
trials, irrespective of how many good ones were found). If a good random
position is found, the counter rests to zero (thus encouraging further
tests). But when even a single good position cannot be found until the
termination limit, it is highly unlikely any other will be found (the
object is too big). So we can simply stop the search.