Naming of datasets in Arithmetic
It often happens that a dataset (in a file) is needed more than once during
an Arithmetic process (for example with the `where' operator). But until
now, it was necessary to give the filename and extension containing the
dataset everytime it was needed. This would cause Arithmetic to read the
same dataset into memory multiple times while also making the command-line
very complex and hard to read.

With this commit a new `set-' operator has been defined that will give a
name to the popped dataset which can then be used multiple times in the
arguments to Arithmetic.

Also, since TIFF files can also have multiple extensions and Arithmetic can
now read them, the part where we check for the number of HDUs now checks
for all files with multiple extensions. To make this possible, the new
`gal_array_name_recognized_multiext' library function was defined.