Corrected usage of named dataset in Arithmetic
A few problems in using a named dataset in Arithmetic are now fixed: 1) in
`operands_add', we weren't checking if `filename!=NULL'. 2) The
`operands_pop' should be completely ignorant to the naming of datasets (it
is irrelevant for popping from a stack). So the check in `operands_pop' was
removed.