Dependency installation with package managers on top webpage
The page describing the commands to install all Gnuastro's dependencies
with a package manager is very convenient for early users. So to help them
find it, a link to that page has been added to the top webpage.

Also, a typo in the NEWS file (for a bug name) was fixed.