BuildProgram's configuration file only built with -I or -L options
By definition, the CPPFLAGS and LDFLAGS environment variables are designed
to pass options to the pre-processor and linker and in particular the `-I'
and `-L' options to specify the directories to look for headers and
libraries.

Until now, when building BuildProgram's configuration file, we just assumed
that only the `-I' option is given to CPPFLAGS and `-L' to LDFLAGS. But
while packaging in Debian's `pdebuild', I noticed that it is passing `-W'
options to these two environment variables: `-Wdate-time
-D_FORTIFY_SOURCE=2' and `-Wl,-z,relro -Wl,-z,now' respectively.

This might happen in other situations is well. With this commit, the script
to add the `includedir' and `linkdir' directories to BuildProgram's
configuration file now only uses strings that start with `-I' or `-L'. Any
other string in these environment variables will be ignored.

This fixes bug #54346.