Arithmetic's where operator checks for blank values
When the condition operand of the `where' operator has blank values, this
operator assumed they are true. With this commit, a check is done on the
condition operand and only when it is not blank *and* non-zero will the
value be modified.

This bug was reported by Dmitry Oparin.

This fixes bug #54358.