Reusing a name in Arithmetic
Until now, if the `set-' operator was given the same name for two datasets,
it would abort, complaining that the given name is already used. However,
in some contexts, you don't know the total number of files to read into
memory, in those cases, it is easier to use the same name many times (while
freeing the dataset that the name previously pointed to).

With this commit, this is done: if `set-' is given a name of an existing
dataset, the old dataset will be freed and name will point to the new one.

Also, the functions in `bin/arithmetic/operands.c' were re-organized to be
easier to read: the operand name functions were separated from the popping
and adding ones.