BuildProgram's configuration file start created with cat
Until now, we would first use `cp' to copy the top of BuildProgram's
configuration file into the build directory, then append the `-I' and `-L'
options to it. But while running `make distcheck' I noticed that the `cp'
will preserve the permissions of the source file and will thus crash the
build when the builder doesn't have write permissions on the source.

So with this commit, instead of using `cp', the top of BuildProgram's
configuration file is read with `cat', then it is redirected to the new
configuration file for the later steps to append to.