Collapse by extrema now uses uint8_t for counter
After its definition, we were using `int32_t' type for the `num' dataset of
collapsing by extrema. But unlike the average (for example), we don't
actually need the number of elements, we just need to know that atleast one
element was present. So like the collapsing by sum, the `num' dataset in
collapsing by extrema now uses `uint8_t' to simply put a `1' where we have
data and 0 where we don't. In this way, we won't consume too much memory
and will be faster.