Statistics check images in output directory if given
Throughout Gnuastro, when an output is specified, the check images are
written in the same directory. However, until now implementing this in the
Statistics program was forgotten! With this commit, Statistics also puts
the check images (for tiles and sky value) in the same directory that
contains the output file (if an output is given).