Proper sanity checks before mode symmetricity calculation
When measuring the symmetricity, we weren't checking if the mode is
actually larger than the `a' (0.05 quantile) point (which is a hidden
assumption). So a check has been added.

Also, since the `MODE_SYM' function works on any type, we were doing
integer division the final step of this function. To be safe, we now cast
the two subtractions to `double' before division.

This fixes bug #54406.