ADS link for Plank 2015 cosmology parameters paper in comments
The most common place to get BibTeX information of papers in astronomy is
ADS (not the separate journal webpages). However, when citing
CosmicCalculator's cosmology, we have only put a link to the main paper's
webpage in A&A. So an ADS link was also added to make it easy for people to
cite the paper that this cosmology comes from.