Spell check on new parts of book
A spell check was done on the new parts of the book prior to the next
official release.

Also, I noticed that in the tutorial's C program on estimating cosmological
distances, we were still using the Plank 2015 results, while in the
previous commit, we moved to Plank 2018 for the default CosmicCalculator
parameters. So that was also corrected.

In a few places, I also corrected the paragraph line length (which was
still set to 70 characters from the early days of Gnuastro, but later
changed to 75).