--enable-debug disables shared libraries also
Until now the configure-time `--enable-debug' option would only disable
optimization and add debugging flags. But when debugging, shared libraries
can also cause a lot of trouble/pain. Therefore in the manual, we had
always also included `--disable-shared' with the `--enable-debug' option.

Since its almost always necessary, the `--enable-debug' option also
disables shared libraries. In the explanations of this option, a
description has been given for the rare occasions that shared libraries are
necessary for the debug.