Input info as keywords in output of NoiseChisel and Segment
It can be very helpful to have all the parameters that produced the output
in the actual output file to allow reproducibility and inspection of
differences (when playing with parameters) later.

To make this, some minor additions needed to be done in the FITS keyword
writing functions of the FITS library. The names of the functions that
write into a `fitsfile' pointer now has a `_in_ptr' suffix to be similar to
the other functions (for example in writing tables and images). Also, two
new functions were defined: `gal_fits_key_list_reverse' and
`gal_fits_key_write_title_in_ptr'.