All programs write their configuration in FITS output
In the previous two commits, a feature was added to NoiseChisel and Segment
to write their input configuration options into the output FITS file. But
those functions were very basic and we had to manually go through all the
options and their names, value pointers and documentation. Maintaining this
list (keeping it in sync with the option updates) was a nightmare: with a
high potential of human error.

But the whole process was basically identical to the --printparams (or
`-P') option. So the `gal_options_as_fits_keywords' was added to the
internal `lib/options.c' library to write all the input values into a
`gal_fits_list_key_t' list at the start of all programs. A new
`gal_fits_key_write_config' function was also created to allow easy writing
of this keyword list into the zero-th extension of the output FITS files.

The old "Output headers" section in the book was also re-written to
describe this new feature, and thus its now called "Output FITS files".