Statistics prepares FITS keywords of configuration when necessary
Until now, Statistics would prepare the input configuration as FITS
keywords in any situtation. But `ui_add_to_single_value' isn't yet prepared
and in those cases, there is no output file any way. So this is an extra
step and would cause a crash.

With this commit, the keywords containing the input configuration are only
written when necessary.