Configuration information in Statistics, only in FITS outputs
Rencently we added a feature for programs to save their configuration in
the first extension of their outputs if they are FITS files. But for
Statistics (that can produce FITS or plain text tables), we weren't
checking if the output is indeed a FITS file or not before this. Therefore
we came up with a CFITSIO error, complaining that the file isn't FITS.

With this commit, a check is added to avoid the crash.