Minor edits in building from package managers
A line was added to the description in `apt-get', that Gnuastro is
available in Debian. The format of the statement for `zipper' was also made
in the same format. Also some of the secondary links in the description of
`zypper' were taken to the footnote to make the text easier to read in PDF.