MakeProfiles only checking validity of column when relevant
When unreasonable values are given (for example a radius equal or smaller
than zero to certain column values MakeProfiles aborts with a notice to the
users. This is a good warning and can help avoid unexpected behavior (for
example the user specifying a column by mistake).

But when the profile doesn't need a parameter, such checks only cause
confusion (for example for a point profile, which never uses the radius,
axis ratio or truncation values). With this commit, checks have been added
for the point profile (which doesn't need some values): when the row's
profile is a point, the value is not checked.

This bug was reported by Michael Stein.

This fixes bug #54526.