NoiseChisel uses outlier rejection algorithm on qthresh
Until now, the user had to manually specify a quantile to remove some tiles
from the quantile threshold interpolation step. But this solution was not
generic and too ugly (too dependent on each particular input dataset).

A new outlier identification algorithm was re-implemented in Gnuastro's
statistics library (the library function had been commented out) in the
previous commit for this purpose. With this commit, it now replaces the
manual quantile identification of NoiseChisel. To configure this new
algorithm, three new options have been added to it as described in the book
and the NEWS file.

This task was found after a great discussion with Pierre-Alain Duc on some
of his datasets during the Lorentz center workshop on "The Bewildering
Nature of Ultra-diffuse Galaxies", in Leiden, The Netherlands.