Mean distance from median used for identifying signal in tiles
To identify the presence of signal in a tile, until now NoiseChisel and
Segment would use the quantile difference of the mode and median. But the
mean's shift is more sensitive to skewness (the presence of signal). It
also has another advantage that it is very easy to measure and will not
fail (unlike the mode finding algorithm which can fail). So with this
commit, in the Statistics and NoiseChisel programs, the mean is now used to
identify the presence of signal in noise.

While working and testing this change, many other issues were fixed and
implemented. Especially in the outlier finding algorithm of the recent
commits. Since the mode isn't used any more, the `--mirrordist' option was
removed from both Segment and NoiseChisel. Also the old `--modmedqdiff' was
renamed to `--meanmedqdiff'.

In the outlier rejection algorithm, the number of prior elements to use for
finding the outlier is now not an option to that function: it starts from
the median and looks half the dataset backwards. This change also
propagated into NoiseChisel and Statistic's options: the `--qthreshoutnum'
option was removed. Also, outlier rejection is now implemented in the
pseudo-detection and final Sky measurements of NoiseChisel. So to be
generic, the options were also renamed to `--outliersclip' and
`--outliersigma'.

While looking into the details of NoiseChisel, I noticed that we weren't
sigma-clipping the tile values prior to writing the Sky
measurements. Therefore the `--sigmaclip' option was redundant. It is now
properly used to reject outlier pixels.

The default `--noerodequant' was increased to 0.99 instead of 0.9331. This
had a major effect in creating bad false positives, so it is now encouraged
to decrease it on a case-by-case basis. The same goes for `--snquant' and
`--detgrowquant'. They were increased to be less prone to false positives
on shallow or mock datasets. They were mainly tuned for deep datasets with
strong correlated noise (which are not too common).

While working on a dataset by Gaspar Galaz (which was inverted and needed
to be subtracted from its maximum), I noticed that the pseudo-detection
step doesn't do Sky subtraction (bug #54579), therefore with this commit,
this step now also accounts for the Sky value.

To make it easy to apply the outlier rejection algorithm in NoiseChisel and
Segment, a new internal library (`tile-internal') was defined in the
library. For the time being, I think its too specific to be defined as a
general library function. But we might do that later if needed.

Given all the new features (outlier removal and using the quantile of the
mean instead of the mode), Gnuastro's default tile size is now set to 30x30
pixels.

Finally, all the relevant parts of the book were updated for these
features. In particular the two tutorials.

Other minor work in this commit:

 - Arithmetic now has a new `--onedasimage' option to save 1D datasets as
   an image, not a table.

 - Fits program now prints `no-data' (instead of the data type) when
   listing the HDUs in a FITS file and there is no data in a HDU.

 - MakeProfiles now writes all the profiles information in the 0th
   extension of the individual profiles it builds. The merged dataset will
   just have the over-all MakeProfiles options.

 - In the library, a new `gal_blank_flag_apply' was added to make it easy
   to mask some tiles.

This fixes bug #54579.