Minor corrections in Segment: warning message and flag setting
Based on a real-world mistake by myself(!), a warning will be printed by
NoiseChisel and Segment if the value to `--snquant' is very small (less
than 0.1). The scenario that this happens if fully described in the comment
above it: mistaking false-detection rate (where lower is better) with
purity (where higher is better).

Also, in Segment, I noticed that while we were intending to set the blank
flag to no flags, we were mistakenly setting it to containing blanks (using
the bit-wise OR operator). This has also been fixed with this commit. So
this commit will allow a small preformance improvement.