No outlier rejection in NoiseChisel's Sky and Sky STD
Until now, we were doing outlier rejection on all three stages that
NoiseChisel needs values over the whole tessellation (the quantile
thresholds, the pseudo-detection Sky and its STD as well as the final Sky
and its STD). But for the latter two it was too much and infact outliers
are infact necessary (and must not be removed). Otherwise we won't be able
to subtract gradients.