Informative message when mandatory dependency is missing
Until now, when a mandatory dependency was missing, there was only a
single-line message which was hard to interpret for new users (especially
people who aren't familiar with the libraries or haven't read the manual
yet). The message would also abort the configure script, so it wasn't
possible to see what other mandatory dependency we are missing.

Since many questions were asked about what the single-line error messages
are, and how to fix them, a more informative message is now printed. It
also lists all the dependencies in one run of `./configure'.