NoiseChisel can ignore blanks in fraction of undetected pixels
Until now, the reference for using a tile for estimating the Sky value was
the number of undetected pixels divided by the total tile area. With this
commit a new `--skyfracnoblank' option is added. When called, the
denominator of the fraction will be the number of non-blank elements within
the tile (not whole tile size). This helps in situtations where the lare
large blank/masked regions scattered across the image, causing a failure in
all tiles.