NoiseChisel's opening connectivity corrected in verbose mode
When printing the steps of the analysis, instead of printing the number
directly, NoiseChisel was actually converting the number to a string, then
printing the string. With this commit, the number is written in the string
directly to avoid extra complications.

Also, to avoid confusion in using the "connectivity" term (for example with
Arithmetic, where connectivity is defined from 1 to the number of
dimensions, for example 4-connected neighbors are have a 1-connectivity),
NoiseChisel now reports things like `4-connected' instead of
`4-connectivity'.