Statistics' ignoreblankinsky implemented in check images also
The `--ignoreblankinsky' option of Statistics (when calculating the Sky
value) wasn't being implemented in the check images while it worked nicely
in the output. This could be confusing. With this commit, it has the same
effect in the check images as the output image.

Also the verbose-mode printed statements are corrected to be more clear and
a statement is also added for the check image.