New NoiseChisel option to consider blank pixels as foreground
Until now blank pixels were treated as foreground in NoiseChisel's binary
processing steps. This would cause many false positives when there were
lots of blank pixels in the dataset. With this commit the default behavior
is to consider blank pixels as background and there is a new option
(`--blankasforeground') if someone wants to have the old behavior.