Extra explanation on CDELT in general tutorial
For people that aren't yet familiar with the WCS keywords the step of
deriving the pixel scales, a more complete explanation on the CDELT keyword
is given in the general (second) tutorial. The text where we used CDELT in
the large object detection tutorial is also edited to be more clear.