Unified GSL random number generator environment setup
Until now the three programs that use GSL's random number generator
(MakeCatalog, MakeProfiles, and MakeNoise) would each call GSL's functions
directly. But this is very inconvenient when an issue is found that must be
checked and fixed in all of them. A wrapper internal function
(`gal_checkset_gsl_rng') has been introduced with this commit to simplify
this process.

This was sparked by finding out we weren't writing the proper seed value
into the FITS keyword of MakeProfile's outputs when `--envseed' was
used. To avoid having to check/fix the problem in all such programs, we now
have this high-level GSL wrapper.