Corrected problem in dependency missing message at configure time
We were mistakenly using the `-L' flag for `CPPFLAGS' in the message that
the configure script prints when a mandatory dependency is missing, causing
the suggested commands to not work. It is corrected in this commit along
with some small edits to make it more clear and easy to read.

This bug was found by Mamta Pommier.