Link to a amateur image of M51, showing its wings in tutorial
While I was looking around in Reddit this evening, I came across a very
nice 12 hour exposure of M51, which clearly shows the very low surface
brightness signal that is detected in the tutorial, but invisible in the
input image. So using this other image is a good demonstration of the
reality of the detection for those who are dubious about the statistics.