Watkins et al 2015 added in tutorial to show reality of detection
In Aaron Watkins' 2015 paper, deep images show the outer wings of M51 at
much higher signal to noise ratios that are visible to the eye in noise. To
further help convince people on the reality of this detection, it is cited
after the detection step and before measuring the surface brightness limit
in the tutorial.

Some other parts of the text in the tutorial were also edited for easier
reading.

This was suggested by Ignacio Trujillo.