Optional dependency status in missing mandatory dependecy message
Until now, if a user was missing a mandatory dependency, the configure
script would abort with a message on which mandatory dependency they are
missing. But since the user goes after installing those dependencies at
this stage, it can also be helpful to them if they know which optional
dependencies they are missing.

With this commit the message printing the mandatory dependencies is moved
after checking optional dependencies. If any mandatory dependency is
missing, it will then print the message before, but it will also print the
status of missing optional dependencies.

Also, following the mandatory dependency crash message, in the final
configure message, a link to the webpage of each missing optional
dependency was added to help the users get started in installing them.