Segment also removes sky clumps in tiles without any usable
In the check image, each step is shown as a separate extension. Segment
first finds all the possible clumps in a large tile, but it doesn't use all
of them: many of the small ones are removed before going to the S/N
phase. Therefore there are two extensions: one for all the clumps, and
another where only clumps that are used for S/N calculations are shown.

I just noticed that when a tile has no usable clumps, none of its clumps
get removed in the check image. This happend because we were setting the
`cltprm.snind' pointer to NULL (when there are no usable clumps) too early!
Also, `clumps_correct_sky_labels_for_check' didn't actually do anything on
tiles with no usable clumps. With this commit both these issues are
fixed. In particular the latter case, just sets all the positive-valued
pixels to `GAL_LABEL_INIT'.

This fixes bug #54782.