New --numhdus option in Fits prints number of extensions/HDUs
This new option will print the number of extensions in a FITS file (and
nothing else). The thought of it actually came up as I was writing a script
myself and noticed the lack of a simple step like this in my script.

I also noticed that we weren't closing the opened FITS file in
`fits_print_extension_info'. So this is also corrected.