The next pointer of the copy of named dataset reset to NULL
When a named operand is copied to be re-used in the stack, Arithmetic
deleted (freed and set to NULL) the `name' pointer, but not the `next'
pointer. This would cause a crash if the copy of the named dataset was
later renamed. With this commit, we also set the `next' pointer to NULL.

This fixes bug #54810.