Random number generator seed with unsigned long type
Until now the type of the random number generator seed was `uint64_t'. But
after testing on a 32-bit system, I noticed that there is no FITS 64-bit
unsigned integer type on those machines, so when we were trying to write
the seed in a FITS file header, the program would crash. The main
complication was that GSL's random number generator seed has `unsigned
long' type (which can differ on 32-bit or 64-bit systems).

With this commit, there is a new type macro called
`GAL_TYPE_ULONG'. Similar to `GAL_TYPE_LONG', it is set and build-time and
is just an alias for the respective fixed-width type. This new Gnuastro
type now enables the use of `unsigned long' for the `rng_seed' variable of
the programs that use random numbers (MakeCatalog, MakeNoise and
MakeProfiles).