Relative extension of range in automatic binning in Statistics lib
In order to count the last element when estimating the bin range
automatically, Until now, in `gal_statistics_regular_bins' we would simply
add 1e-6 to the maximum value. This would cause problems when the range of
the dataset is smaller than that (for example when using angestroms). Using
a very small value was also not possible, because of floating point errors.

So now, the histogram function uses a different check when it is
considering the last bin. This fixes all the complications above.