Statistics can read from Standard input
Until now, Gnuastro's programs would only take their input from a
file. However, atleast for plain text files, it is very annoying (and prone
to bugs) to be forced to create a new temporary file and break a series of
pipes.

One of the most common cases was the Statistics program: when you make a
column with something like `awk' and want to want to do some basic
statistical measurements on it. So to build the necessary infra-structure
(and use in all the programs), the Statistics program was chosen to start
with.

With this commit, an initial implementation of this system has been added
to Statistics, it will later be extended to any other program that takes
plain text files as input.